session17/library/requirements.txt
django
django-bootstrap4
djangorestframework
gunicorn